requirements.txt
astpretty==3.0.0
astunparse==1.6.3
cmake==4.0.3
colorlog==6.9.0
comm==0.2.3
dnspython==2.7.0
expecttest==0.3.0
fsspec==2025.7.0
func_timeout==4.3.5
gensim==4.3.3
graphviz==0.21
hf-xet==1.1.8
hityper==1.0.3
huggingface-hub==0.34.4
hypothesis==6.137.1
ipywidgets==8.1.7
jupyterlab_widgets==3.0.15
Levenshtein==0.27.1
libcst==0.4.2
lintrunner==0.12.7
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
optree==0.17.0
pandas==2.3.1
pillow==11.0.0
PyCG==0.0.8
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-etcd==0.4.5
RapidFuzz==3.13.0
regex==2025.7.34
safetensors==0.6.2
scipy==1.13.1
six==1.17.0
smart_open==7.3.0.post1
sortedcontainers==2.4.0
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.21.4
torchaudio==2.8.0
torchelastic==0.2.2
torchvision==0.23.0
transformers==4.55.2
triton==3.4.0
types-dataclasses==0.6.6
tzdata==2025.2
widgetsnbextension==4.0.14
wrapt==1.17.3
zstandard==0.23.0
GitPython
dateparser
datasets
scikit-learn==1.6.1